requirements.txt
Django==1.10.4
django-modeltranslation==0.12
django-parler-rest==1.4.2
django-webpack-loader==0.4.1
djangorestframework==3.5.3
mysqlclient==1.3.9
olefile==0.43
Pillow==4.0.0
pngquant==1.0.5
